Hole in one
Got this on my first try. I'd hoped it might be secretly trickier,
but it's just the usual reference cycle pains that blocks present.

This solution is cleaner than that in the blog post
(http://sealedabstract.com/code/nsnotificationcenter-with-blocks-considered-harmful/)
in that it does not have to dance around the `self` reference or resort to
direct ivar access.